Large commit: Contains the .jpa.dao.* and .jpa.bean.* classes for handling the data layer.
Contains the DAO classes and the Entity Beans for the JPA based data
layer.  Also provided are the required Hibernate jar files (you must use
these versions for the code to compile).  In the web directory, there is
a test JSP (/test/testDAO.jsp) which provides a test function and sample
code.